% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFiles}
\alias{getFiles}
\title{Retrieve theme files}
\usage{
getFiles(fn, dn = "theme", pkg = "stationery")
}
\arguments{
\item{fn}{Vector of file names}

\item{dn}{"theme", or directory name used in package inst for
theme information.}

\item{pkg}{Package name, default "stationery"}
}
\value{
TRUE if succeeded
}
\description{
A wrapper for file.copy that retieves files from the dn = "theme"
directory of a package
}
\author{
Paul Johnson<pauljohn@ku.edu>
}
